feat(update_psrcat): create pulsar catalog that can be used by FluxCatalog
This adds more flux measurements for the pulsars and fits the measured
flux to a model so that we can predict fluxes for the pulsars
in the CHIME band (and display the catalog in FluxCatalog).  It
also adds the `loaded_collections` method to FluxCatalog that returns
the collections that are currently loaded.  Finally it fixes a warning
that was being thrown by np.vstack.