docs(layout): build docs for members imported from _db_tables
autodoc ignores imported members. layout wants to pretend that these
imported members are its own, though. If :imported-members: is set to
True to tell autodoc to generate documentation for these, we get lots of
errors wherever we import somethoing with "from * import *". So the
docstring now bends links to these imported members as a solution.