doc(layout): Encourage use of username with set_user
The previous layout docuemntation implied you needed to pass
your "real name" (whatever that is) when calling `set_user`.
This is problematic for a few reasons:
- it confuses the notion of someone's actual name and the
  value of the CHIMEwiki user_real_name field (which is what's
  being matched here)
- the user_real_name field is typically populated by sysadmins
  when creating a wiki account.  It's available and can be
  changed by users in the wiki profile page, but most wiki users
  don't change it nor do they even know what the value has
  been set to.
- There's absolutely no validation performed on the column.  I've
  seen misspelled names, blank names, email addresses in this
  column.
- There's no requirement that the values of `user_real_name` be
  unique.

This changes the docs to encourage using CHIMEwiki usernames when
calling `set_user`.  Usernames are much better to use because:
- they're unambiguous in form, mean simple string matching is a
  reasonable thing to do.
- they're well known to users, since users use them to log into the
  wiki

Note: this doesn't change any functionality: users are still
able to pass the value of the user_real_name field associated
with their account to this function.