chore: Delete old files
We don't support HKP data files anymore, so there's no
more need for examples on how to use them.

The SQL scrips are ancient (>10 years old).  They're woefully
out-of-date.  These days, if we were to re-create a data table,
and didn't want to do it via peewee, we'd pull the CREATE TABLE
statement out of the nightly DB export.