fix(hfbcat): Re-initialise class variables
Because it subclasses from FluxCatalog, by default will inherit
the class variables from FluxCatalog, but that's where FluxCatalog
stores the catalogs it loads at import time.  (Which seems like a
weird, an un-Pythonic, effect of importing a module.)

The side effect is that all the FluxCatalog catalogs end up in the
HFBCatalog as well, which I suspect is not intended.  So here's
a fix to re-initialise these class variables to nothing, so that the
FluxCatalog catalogs don't pollute the HFBCatalog.

Now, only the catalogue(s) loaded explicitly into HFBCatalog will
exist in it.