fix(alpenhorn_holography): Moved from ch_alpenhorn
This is a transfer of the "alpenhorn_holography" script
that used to be in `ch_alpenhorn`.

Despite the name, this script is unrelated to alpenhorn.  I need to
move it out of `ch_alpenhorn` so that repo can be retooled for
alpenhorn-2.

Note: though I've renamed the module itself to `ch_util.holography_cli`,
the installed script is still called `alpenhorn_holography` to avoid
making breaking changes.

I've done some light editing to the script for the deprecation of
`ch_util.ephemeris` and to get the code to conform to the required linting
standard in this repository.